Move auth into databricks/auth since it's exclusively related to Databricks.
