+++ REFACTORING_SUMMARY.md
@@ -11,7 +11,7 @@ This refactoring successfully split the `VolumeHandler` class into two distinct 
 
 ### Volume (Pydantic Model)
 - **Purpose**: Pure data model for volume identity (catalog, schema_name, volume_name)
-- **Features**: 
+- **Features**:
   - No I/O operations or external dependencies
   - Methods for URI construction: `get_full_name()`, `get_volume_path()`, `get_file_path()`
   - Pydantic validation and type safety
@@ -19,7 +19,7 @@ This refactoring successfully split the `VolumeHandler` class into two distinct 
 ### VolumeFileStore
 - **Purpose**: Handles all file operations for a specific volume
 - **Constructor**: Takes `Volume` and `WorkspaceClient` as explicit dependencies
-- **Methods**: 
+- **Methods**:
   - `volume_exists()` - Check if volume exists
   - `list_files()` - List files in volume
   - `file_exists()` - Check if specific file exists
@@ -56,12 +56,12 @@ The original `VolumeHandler` class is maintained as a wrapper around the new com
 ### New Approach (Recommended)
 ```python
 from backend.databricks.volume import Volume, VolumeFileStore
-from backend.auth import get_databricks_auth
+from backend.databricks.auth import get_databricks_auth
 
 # Create volume identity
 volume = Volume(
     catalog="my_catalog",
-    schema_name="my_schema", 
+    schema_name="my_schema",
     volume_name="my_volume"
 )
